requirements.txt
beautifulsoup4==4.11.1
Django==3.2
django-jquery==3.1.0
django-pgviews==0.5.7
gunicorn==20.1.0
olefile==0.46
Pillow==9.2.0
psycopg2-binary==2.9.4
python-decouple==3.6
python-docx==0.8.6